Refactor: update tsconfigs and package.json to reflect monorepo
