+++ packages/core/tsconfig.json
@@ -1,0 +1,9 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+        "outDir": "./lib",
+        "rootDir": "./src"  
+    },
+    "include": ["src/**/*.ts"],
+    "exclude": ["node_modules", "lib"]
+}
