+++ packages/mcp/cli.js
