+++ packages/mcp/index.d.ts
