+++ packages/mcp/index.js
